Fix for Issue #2 Exception when using a package without any components defined
Description: 
If there are no components defined in a package the application crashes
randomly, as soon as we try to read text from a binary file.
This is due to the fact, that if no components are defined we iterate
over the whole base directory of the choosen package.

Resolution:
We no more iterate over the base directory if no components are
specified.
We just skip component parsing instead.