Fix for Issue#3: Finished Message
Description:
At the moment it is not visible when the launcher has finished.
I am going to add a message box after clicking its 'OK' button the
'POST_CMD' will be executed.

Resolution:
Message boxes now appear as soon as the launcher is about to terminate.
POST_CMD is executed after the message box is closed.