Implementation Issue #12 - Log stdout from new process
Description:
POST_CMD process output is currently not seen.
Log it as all other output to the display and the log file.

Resolution:
POST_CMD now executed on another thread so the launcher
will not freeze.
Log stdout from subprocess cancelled.
This would lead to the launcher not being able to close after
starting POST_CMD process.
The process started should take care itself about logging.
If a shell script or kinda things shall be executed it would be better
to redirect the output from within the script.