Implementation for issue #29 - Preview: Deselect components which should not be downloaded
Description:
In preview it should be possible to deselect components which should not
be downloaded.
This makes it possible to download config files at first time and
afterwards deselect them to not
overwrite local configs.
Additionally it should be possible to flag components as "required" so
some components can not be deselected.

Resolution:
"required" is now a valid attribute for the component configuration.
Default value is "true", if you want the component to be optional
set the attribute to "false".
The preview reflects this as checkbox.